Cookie

+++ views/index.tpl
@@ -10,10 +10,16 @@
         <div class="col">
           <div class="card__header">
             <h4 class="title-4"><?=$value['title']?></h4>
-            <div>
-              <a href="edit.php?id=<?=$value['id']?>" class="button button--edit">Редактировать</a>
-              <a href="?action=delete&id=<?=$value['id']?>" class="button button--delete">Удалить</a>
-            </div>
+            
+            <?php if(isAdmin()) { ?> 
+
+              <div>
+                <a href="edit.php?id=<?=$value['id']?>" class="button button--edit">Редактировать</a>
+                <a href="?action=delete&id=<?=$value['id']?>" class="button button--delete">Удалить</a>
+              </div>
+
+            <?php } ?>
+
           </div> 
           <div class="badge"><?=$value['genre']?></div>
           <div class="badge"><?=$value['year']?></div>
